requirements.txt
# Production Dependencies
# Sentiment Analyzer - Demo 1 (001-interactive-dashboard-demo)
#
# For On-Call Engineers:
# - If Lambda fails with ImportError, check these versions match deployed layer
# - boto3 version must be compatible with Lambda runtime (Python 3.13)
#
# For Developers:
# - Pin exact versions for reproducible builds
# - Update versions only with explicit testing
# - See requirements-dev.txt for test dependencies

# AWS SDK - Lambda runtime includes boto3, but we pin for local testing
boto3==1.42.5
botocore==1.42.5

# ML/NLP - DistilBERT sentiment analysis
# Note: torch is large (~2GB); Lambda layer built separately
# See infrastructure/scripts/build-model-layer.sh
transformers==4.57.3
torch==2.9.1

# Web Framework - Dashboard Lambda
# FastAPI for endpoints, Mangum for Lambda adapter
fastapi==0.124.0
mangum==0.19.0
uvicorn==0.38.0  # Local development only
sse-starlette==3.0.3  # Server-Sent Events for real-time updates

# HTTP Client - Financial API adapters (Tiingo, Finnhub)
requests==2.32.5
httpx==0.28.1  # Async HTTP client for financial APIs

# Data Validation - Input/output schemas
pydantic==2.12.5
email-validator==2.3.0  # Required for Pydantic EmailStr type

# Logging - Structured JSON logs for CloudWatch
python-json-logger==4.0.0

# AWS X-Ray - Distributed tracing (Feature 006)
aws-xray-sdk==2.15.0

# Email Notifications - SendGrid (Feature 006)
sendgrid==6.12.5

# AWS Lambda Powertools - Enhanced observability
aws-lambda-powertools==3.23.0

# JWT Authentication - Feature 075 (validation gaps)
PyJWT==2.10.1
